requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.1.1
django-cors-headers==3.13.0
django-money==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
py-moneyed==2.0
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
typing-extensions==4.3.0
tzdata==2022.4
